backend-django/core/oauth/oauth_schema.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
"""
OAuth Schema - OAuth 数据模型
定义 OAuth 相关的请求和响应数据结构
"""
from typing import Optional
from ninja import Schema


class OAuthCallbackSchema(Schema):
    """OAuth 回调请求参数 (通用)"""
    code: str
    state: Optional[str] = None


# 保留旧的 Schema 名称以兼容
GiteeCallbackSchema = OAuthCallbackSchema


class OAuthLoginResponseSchema(Schema):
    """OAuth 登录响应"""
    access_token: str
    refresh_token: str
    expire: int
    user_info: dict


class GiteeUserInfoSchema(Schema):
    """Gitee 用户信息"""
    id: int
    login: str
    name: str
    avatar_url: str
    email: Optional[str] = None
    bio: Optional[str] = None
    blog: Optional[str] = None
    company: Optional[str] = None
    location: Optional[str] = None
